kernel: cast NVIC_IRQ_COUNT for the usize compare in the rp2040 usable() gate
The ARMv6-M interrupt-controller `usable()` predicate added in the #112
interrupt-allocator work compared `cpu_int as usize` against
`soc_rp2040::NVIC_IRQ_COUNT`, which is a `u8`. The Xtensa host build
never compiles this arm, so it only surfaced under `make check-all`'s
thumbv6m pass.

Farhan Husain <cooljackal@gmail.com>

+++ kernel/src/interrupt_controller.rs
@@ -74,7 +74,7 @@ impl InterruptController for Selected {
         // can only ever land on `cpu_int == source`; `route` is what enforces
         // that. Every existing vector is a candidate here and the route gate
         // rejects the mismatches.
-        (cpu_int as usize) < soc_rp2040::NVIC_IRQ_COUNT
+        (cpu_int as usize) < soc_rp2040::NVIC_IRQ_COUNT as usize
     }
 
     unsafe fn clear_pending(mask: u32) {
